Code divided into individual files
Separated the code so that later after testing, this can all be converted to an executable program.
